Linux compiler warnings are now gone :)

+++ src/cgame/cg_fireteams.cpp
@@ -22,7 +22,7 @@ panel_button_t fireteamTopBorder =
 	NULL,
 	NULL,
 	{ 10,                     129,        204,       136 },
-	{ 1,                      255 * .5f,  255 * .5f, 255 * .5f, 255 * .5f, 1, 0, 0},
+	{ 1,                      (int)(255 * .5f),  (int)(255 * .5f), (int)(255 * .5f), (int)(255 * .5f), 1, 0, 0},
 	NULL,                     /* font		*/
 	NULL,                     /* keyDown	*/
 	NULL,                     /* keyUp	*/
@@ -35,7 +35,7 @@ panel_button_t fireteamTopBorderBack =
 	"white",
 	NULL,
 	{ 11,                     130,202, 134 },
-	{ 1,                      0,  0,   0, 255 * 0.75f, 0, 0, 0},
+	{ 1,                      0,  0,   0, (int)(255 * 0.75f), 0, 0, 0},
 	NULL,                     /* font		*/
 	NULL,                     /* keyDown	*/
 	NULL,                     /* keyUp	*/
